DRILL-7969: Add support for reading and writing Parquet files using Brotli, LZ4 and Zstandard codecs (#2321)
Adds support for all the standardised Parquet compression codecs beyond GZip
and Snappy by making use of the airlift/aircompressor library with a fallback
to parquet-mr for compression for codecs not implemented in parquet-mr.

A new, delegating CompressionCodecFactory implementation is included.  This
handles the routing of (de)compression to the correct lib while having a
minimal impact on the calling code in the Parquet reading and writing parts
of the Drill codebase.

Add read/write tests and data files for new Parquet codecs.

Exclude and shade aircompressor in other modules.

Don't package or test Brotli compression except for Linux and Mac on AMD64.